requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
langchain-text-splitters
qdrant-client
sentence-transformers
redis
google-generativeai